server shutdown works again...
I'm not sure why, but the server shutdown-er hadn't been working, so I fixed
it, with a `while True: pass` inside of the try so that it doesn't think that
there isn't anything left to do and passes out of the try into NOTHINGNESS.

Basically, this keeps the program inside of the try block until there's a
KeyboardInterrupt (or SystemExit), where it will than correctly kill
everything.

This is a good thing.